+++ jetpack_vendor/automattic/jetpack-post-list/CHANGELOG.md
@@ -11,6 +11,7 @@ This is an alpha version! The changes listed here are not final.
 
 ### Changed
 - Update dependencies.
+- Update package dependencies.
 
 ## [0.8.18] - 2025-06-05
 ### Changed